git clone: use ssh url for private repos
Private repos still come through with an https:// url, which is no
good cause our client only uses a private key for authentication and
would be forced to provide the password on each access.  So, now we
use the ssh_url: git@github.com:owner/repo.git for private repos.

Also, string the last two promises so that a failure of the
backhub.inject() call will still trigger the .fail() method.